Create README.md
This repository will be jam packed with everything you need to know for students, teachers/educators and families. International and global learning journeys can be a scary but new & exciting lifetime experience and I would love to spread my knowledge through experience and resources that helped me get the very most out of my Global Volunteer and Work/Study Abroad opportunities. ~~~"It's not as scary as one may think- until we give it a go, then we'll just never know!"~~~